Seq now carries a list_t<doc_t> -- a 1.2M file still gets processed in 2 seconds
git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1236 581db07f-6e62-4d40-9733-d432c4eebb07

+++ tests/smlng/tag_elim.h
@@ -22,7 +22,7 @@ extern struct SynthS {
 };
 extern union Synth {
   struct SynthS s;
-  unsigned int i;
+  unsigned long long i;
 };
 typedef union Synth synth_t; // not a pointer
 extern $(doc_t, synth_t) up_opt(Buffer::buf_t, doc_t);
